Updated CSS
Removed Media Queries which were there to set the qr-code-container to the centre at different incremental height and width points.

Added min-height: 100vh; to body selector in style.css file. This fixes the issue of the qr-code-container not automatically centring at different viewport sizes, such as from desktop to mobile.