Updated CSS file
Removed fixed height property on qr-code_container and qr-code-image selectors. Padding added to space out text from the image proportionately to go along with the design.

Changes to font-family and font sizing as well as image properties to centre the image and changes to the box-shadow on the image.

Removed media query which was in place to hide overflow on the y axis to prevent the ability to scroll up and down. This was placed in the body selector, overflow on the x axis set to hidden was also added.

Flexbox added with flex direction set to column in the qr-code_container selector for alignment of children.